build(deps): Update versions & pin lodash
There were a bunch of compromised dependencies, so I thought I'd
run through a maintenance update. I've also had to pin
lodash as 'webpack-manifest-plugin' is in the process
of ripping the full lodash dependency out in favour of taking
modular dependencies, but that leaves a P1 vuln in one of its
dependencies. That has now been pinned (caret range still) to
a patched version.